refactor: improve ticket CSV import architecture and fix stale closure bug
- Move validateTicketImportData and processTicketRows from server actions
  (ticketImportActions.ts) to client-side utility (lib/ticketImportUtils.ts)
  to avoid unnecessary network round-trips for pure data transformations
- processTicketRows is now synchronous (returns IProcessedTicketData[]
  instead of Promise)
- Fix stale closure bug in handleProceedFromPreview by adding handleImport
  to its useCallback dependency array and reordering so it's defined after
  handleImport
- Remove redundant useEffect reset block — parent already conditionally
  renders the dialog, so unmount/remount handles cleanup; default board
  selection moved to useState initializer
- Remove dead board resolution code (unmatchedBoards state, boardResolutions,
  handleBoardResolutionChange, board resolution UI section) — all tickets
  use the selected default board
- Remove subcategory from CSV template, column mapping, and resolution flow
- Skip 'board' and 'subcategory' fields during auto-mapping and manual
  column mapping
- Replace raw radio button markup with RadioGroup component in all
  resolution sections for UI consistency
- Add visual indicators (AlertTriangle, Check icons, destructive styling)
  for unmapped required fields in column mapping step
- Simplify board selection step copy now that per-row board override is
  removed
